fixed issue #6
Hi

Line 175 of ZIF_MOCKA_MOCKER~GENERATE_MOCKUP checks the visibility field
for a 0 to determine if the attribute is a private one or not.

On our release of SAP this value is an 'I'.

I've changed my code from

IF -visibility = 'I'."no mocking of private attributes

to

IF -visibility = 'I' or '0'."no mocking of private attributes

Thanx

Nath