Fix issue with mocking abstract classes
When mocking an abstract class, if you don't mock all the abstract methods explicitly you get an error saying that a subclass of an abstract class must implement all the abstract methods of the super class. This gets around that issue by redefining any methods that are abstract even if they don't have registered call patterns.